Updated s3server Project: mero  a73d1ccf94c9469c16a1db1c815bce3c69609d42
MERO-2377 Pool: Some pool versions are having more than 'K' device failures

Pool machine of some versions are having more than 'K' failures when only 'K'
devices are failed. Due to this repair won't be completed.
The cause of this issue is when a device comes online after rebalance device is
removed from pool failed list, but the same device is present in other pool
versions too.
While decrementing the number of failed devices in a pool version, we check
for device to be present in failed list, so this decrement is done only for
first pool version.

This is fixed by decrementing the counter whenever device is rebalanced.

Change-Id: Ibb23d441133af493a805200102a2976fe22e005f