MINT-1138: S3 - S3 commands will fail if user runs nginx in some other port than 80
S3 Apis will fail due to authentication error when nginx is run on
some port other than 80, its because when port number on which nginx
is running is provided by client, nginx strips this port number and send
only host name to s3server. This results in Authentication signature mismatch.
Made change in nginx conf file to use '$http_host' instead of '$host' which
will not strip the port number.

Change-Id: Ia23b4b2daefeec6404bd1eea25bbd5948d73803c