EOS-13325: Enhance jenkins script to take additional arguments to save jenkins job logs.
We need s3server and possibly other component logs for debugging failures seen on jenkins job run.
For that, we must save support bundle before ending the job.
With these changes, we will create support bundle at the end of the jenkins build job and save it to
given location.

Signed-off-by: Amit Kumar <amit.kumar@seagate.com>