EOS-14640: Fix for setting the NextMarker correctly.
The issue is NextMarker was set to the last key of the result set/iteration, before running into iteration/loop.
Due to this, if max_keys=1000 condition hits, the NextMarker was set to the last key in the result set. This is incorrect.
So, setting the last key to key just before breaking the loop/iteration.

Signed-off-by: ‘Dattaprasad <dattaprasad.govekar@seagate.com>